+++ migrations/20180410162630-create-chat.js
@@ -6,7 +6,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INT(11)
+        type: Sequelize.INTEGER(11)
       },
       label_chat: {
         type: Sequelize.VARCHAR(255),
