+++ models/possesses.js
@@ -1,8 +1,8 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   var Possesses = sequelize.define('Possesses', {
-    id_user: DataTypes.INT(11),
-    id_role: DataTypes.INT(11)
+    id_user: DataTypes.INTEGER(11),
+    id_role: DataTypes.INTEGER(11)
   }, {});
   Possesses.associate = function(models) {
     models.Possesses.belongsTo(models.Users, {
